Ventanas
Se separan las ventanas y se realiza cambio en el código de estas
